Refactor & add comments. Split out controller as separate module.
